modified go.mod and moved to dockerfile
